parser/requirements.txt
requests==2.32.0
beautifulsoup4==4.12.3
lxml==5.3.0
anthropic==0.40.0
python-dotenv==1.0.0
